azure-functions/DailyOutboundCycle/__init__.py
"""
Daily Outbound Cycle (v10) — Azure Function Timer Trigger
Runs at 7:00 AM EST (12:00 UTC) every day.

Original: daily_cron_v10.py (Perplexity Computer scheduler)
Migration: Import your existing daily_cron_v10.py logic here.
"""
import logging
import azure.functions as func

# TODO: Import your existing daily cron logic
# from outbound_machine.daily_cron_v10 import run_daily_cycle

def main(timer: func.TimerRequest) -> None:
    logging.info("Daily Outbound Cycle (v10) triggered")

    if timer.past_due:
        logging.warning("Timer is past due — running anyway")

    try:
        # TODO: Replace with actual daily cron logic
        # run_daily_cycle()
        logging.info("Daily Outbound Cycle completed successfully")
    except Exception as e:
        logging.error(f"Daily Outbound Cycle failed: {e}")
        raise
